+++ packages/create-exo-app/templates/minimal/src/main.ts
@@ -8,7 +8,7 @@ const app = new Application({
     width: 800,
     height: 600,
   },
-  clearColor: Color.cornflowerBlue,
+  clearColor: new Color(0x6495ed),
 });
 
 document.body.append(app.canvas);